+++ packages/tts-ggml/src/tts.ts
@@ -4,7 +4,7 @@ import {
 } from "./lib/error";
 
 export interface TTSConfigurationParams {
-  [key: string]: string | number | boolean | undefined;
+  [key: string]: string | number | boolean | string[] | undefined;
 }
 
 export interface TTSJobData {