+++ packages/tts-ggml/tts.d.ts
@@ -1,5 +1,5 @@
 export interface TTSConfigurationParams {
-    [key: string]: string | number | boolean | undefined;
+    [key: string]: string | number | boolean | string[] | undefined;
 }
 export interface TTSJobData {
     type: string;